+++ akshare/__init__.py
@@ -2476,9 +2476,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.10.59 add: add fund_announcement_personnel_em interface
 1.10.60 add: add macro_usa_cpi_yoy interface
 1.10.61 fix: fix macro_cnbs interface
+1.10.62 fix: fix index_hist_sw interface
 """
 
-__version__ = "1.10.61"
+__version__ = "1.10.62"
 __author__ = "AKFamily"
 
 import sys