+++ docs/changelog.md
@@ -63,6 +63,10 @@
 
 ## 更新说明详情
 
+1.10.62 fix: fix index_hist_sw interface
+
+    1. 修复 index_hist_sw 接口
+
 1.10.61 fix: fix macro_cnbs interface
 
     1. 修复 macro_cnbs 接口
@@ -2657,6 +2661,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.10.62 fix: fix index_hist_sw interface
 
 1.10.61 fix: fix macro_cnbs interface